Cleanup, move abstraction_rewriter to separate file
